Add support for MIRA log-linear weight tuning

+++ src/SIL.Core/Progress/IProgressIndicator.cs
@@ -4,7 +4,7 @@ namespace SIL.Progress
 {
 	public interface IProgressIndicator
 	{
-		int PercentCompleted { get; set; }
+		double PercentCompleted { get; set; }
 		void Finish();
 		void Initialize();
 		void IndicateUnknownProgress();
